Add common TS types package.
