+++ packages/components/package.json
@@ -48,14 +48,12 @@
         "@redhat-cloud-services/frontend-components-utilities": ">=3.0.0",
         "sanitize-html": "^2.3.2",
         "@scalprum/core": "^0.1.1",
-        "@scalprum/react-core": "^0.1.7"
+        "@scalprum/react-core": "^0.1.7",
+        "@redhat-cloud-services/types": "0.0.1"
     },
     "devDependencies": {
         "@patternfly/patternfly": ">=4.102.2",
         "node-sass-package-importer": "^5.3.2",
         "@types/react": "^16.9.34"
-    },
-    "optionalDependencies": {
-        "@redhat-cloud-services/rbac-client": "^1.0.104"
     }
 }
