+++ packages/components/src/useChrome/index.ts
@@ -1,3 +1,3 @@
+export * from './useChrome';
 export { default } from './useChrome';
 export { default as useChrome } from './useChrome';
-export * from './useChrome';
