+++ packages/types/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "@redhat-cloud-services/types",
+    "description": "TypeScript definitions for @redhat-cloud-services common typings",
+    "version": "0.0.1",
+    "main": "",
+    "types": "index.d.ts",
+    "publishConfig": {
+        "access": "public"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/RedHatInsights/frontend-components.git",
+        "directory": "packages/types"
+    },
+    "scripts": {},
+    "devDependencies": {
+        "@patternfly/quickstarts": "^1.4.1-rc.1",
+        "@redhat-cloud-services/rbac-client": "^1.0.104",
+        "history": "^4.0.0"
+    },
+    "author": "",
+    "license": "Apache-2.0"
+}
